+++ src/main/java/com/joyous/festivalolle/member/controller/MemberController.java
@@ -82,47 +82,7 @@ public class MemberController {
 	public String join(HttpSession session) {
 		return view_pos + "join";
 	}
-	
-	//id 찾기
-		@GetMapping(value="/findid")
-		public String getfindId(HttpSession session) {
-			return view_pos + "findid";
-		}
-		
-		@PostMapping(value="/findid")
-		public String postfindId(HttpSession session) {
-			memberService.findId(view_pos, view_pos);
-			return view_pos + "findidinfo";
-		}
 
-	//id 확인
-	@GetMapping(value="/idinfo")
-	public String findIdInfo(HttpSession session) {
-		return view_pos + "findidinfo";
-		}
 
-	//비밀번호 찾기
-	@GetMapping(value="/findpw")
-	public String getFindPw(HttpSession session) {
-		return view_pos + "findpw";
-	}
-	
-	@PostMapping(value="/findpw")
-	public String postFindPw(HttpSession session) {
-		memberService.findPw(view_pos, view_pos);
-		return view_pos + "findpw";
-	}
-	
-	//비밀번호 변경
-	@GetMapping(value="/pwinfo")
-	public String findPwInfo(HttpSession session) {
-		return view_pos + "findpwinfo";
-	}
-	
-	@PostMapping(value="/pwinfo")
-		public String postfindIdInfo(HttpSession session) {
-			memberService.updatePassword(view_pos, view_pos);
-			return view_pos + "findpwinfo";
-	}
 }
